requirements.txt

certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.5.0
pytest==8.3.5
PyYAML==6.0.2
requests==2.32.3
urllib3==2.4.0

dotenv~=0.9.9
python-dotenv~=1.1.0

setuptools~=79.0.1
loguru~=0.7.3